+++ lib/src/io/env_parser.dart
@@ -320,10 +320,10 @@ Please, set [dart:output_dir:] to use any dart generation capability.
 
 const _kSampleConfig = '''
 ## output dir for json translations by locale
-output_json_dir: data/output/assets/i18n
+output_json_dir: assets/i18n
 
 ## main entry file to generate the unique translation json.
-entry_file: data/entry/sample.yaml
+entry_file: strings/sample.yaml
 
 ## pattern to applies final variables in the generated json/dart Strings.
 ## Enclose * in the pattern you need.
@@ -339,13 +339,13 @@ intl:
 
 dart:
   ## Output dir for dart files
-  output_dir: data/output/src/i18n
+  output_dir: lib/i18n
 
   ## Translation Key class and filename reference
-  keys_id: Keys
+  keys_id: TKeys
 
   ## Translations map class an filename reference.
-  translations_id: Translations
+  translations_id: TData
 
   ## translations as dart files Maps (available in translations.dart).
   use_maps: false
@@ -372,8 +372,8 @@ gsheets:
 
   credentials_path:
 
-  ## Open your google sheet and get it from the url:
-  ## https://docs.google.com/spreadsheets/d/{ID}
+  ## Open your google sheet and copy the SHEET_ID from the url:
+  ## https://docs.google.com/spreadsheets/d/{SHEET_ID}/edit#gid=0
   spreadsheet_id:
 
   ## The spreadsheet "table" where your translation will live.